predict StackSystem GetVerbsEvent
